+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All changes are backwards-incompatible since it has not reached `1.0.0`.
 
 ### Added
 
+- `StoryInterface` special passage
 - `stringify` utility function.
 
 ## [0.2.1] - 2024-11-27