+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All changes are backwards-incompatible since it has not reached `1.0.0`.
 
 - Saves can now be imported and exported as JSON files.
 - Trying to read an unset story variable triggers a "Did you mean...?" warning in the console.
+- Functions can be used as dynamic attributes for event listener attributes such as `onclick`.
 - `StoryInterface` special passage.
 - `stringify` utility function.
 - `historyLength` configuration option.